One additional test case added for Deck. DeckTest is passing all three testcases.
Deck now has a method for genereating cards. But rather then generating a custom number of cards (as first planned) it can only generate a set* (or several) of 52 cards. Also a method for drawing a card is implemented, but not random. In fact it only pulls the first element of the arraylist at this time.

Deck implements it's own toString now, displaying the content of the
Deck/arraylist.

*A set of cards, referreing two a complete set - 52 cards, not the data
structure.