GameStateMachine: Added testcases for "All in" state Game: added reminder that endRound() is incomplete CardGame: added "putSpecificCard(Card card)" method CardGameTest: added tests for "putSpecificCard(Card card)" Deck: added "removeSpecificCard(Card card)" method DeckTest: added tests for "removeSpecificCard(Card card)"
GENERAL WARNING: I noticed that we have a lot of methods returning the actual private arrays. This is a major security issue, and we should split these methods up so that one version of the method is private, and the other version of the method is public, but return a copy of private attributes.